+++ CAnEn/include/Functions.h
@@ -220,10 +220,22 @@ namespace Functions {
     /**
      * Collapse the time and lead time dimensions of a forecasts and convert
      * them to observations;
-     * @param  Observations
-     * @param  Forecasts
+     * @param  Observations Observations to store results
+     * @param  Forecasts Forecasts to be collapsed
      */
     void collapseLeadTimes(Observations &, const Forecasts &);
+
+    /**
+     * Unwrap the times series of observations to reconstruct forecast times and lead times.
+     * This is the reverse process of `collapseLeadTimes`.
+     * @param Forecasts Forecasts to store results
+     * @param times Forecast times to reconstruct
+     * @param flts Forecast lead times to reconstruct
+     * @param Observations Observations to be unwrapped
+     * @param permissive Whether to allow not-found
+     */
+    void unwrapTimeSeries(Forecasts &, const Times &, const Times &,
+            const Observations &, bool permissive = false);
 
     /**
      * Functions to calculate indices to split a consecutive vector.