+++ apps/anen_netcdf/anen_netcdf.cpp
@@ -53,6 +53,7 @@ void runAnEnNcdf(
         bool overwrite,
         bool profile,
         bool save_tests,
+        bool unwrap_obs,
         bool convert_wind,
         const string & u_name,
         const string & v_name,
@@ -299,11 +300,23 @@ void runAnEnNcdf(
             // Create test observations
             ObservationsPointer test_observations(observations.getParameters(), observations.getStations(), test_obs_times);
 
-            // Copy subset values from original observations
-            observations.subset(test_observations);
+            if (unwrap_obs) {
+                if (config.verbose >= Verbose::Progress) {
+                    cout << "Saving the unwrapped observations as forecasts ..." << endl;
+                }
 
-            // Save subset observations
-            anen_write.writeObservations(fileout, test_observations, false, true);
+                ForecastsPointer unwrapped_observations;
+                Functions::unwrapTimeSeries(unwrapped_observations, test_times, forecasts.getFLTs(), test_observations, true);
+
+                anen_write.writeForecasts(fileout, unwrapped_observations, false, true, "AlignedObservations");
+
+            } else {
+                // Copy subset values from original observations
+                observations.subset(test_observations);
+
+                // Save subset observations
+                anen_write.writeObservations(fileout, test_observations, false, true);
+            }
         }
 
         profiler.log_time_session("Writing forecasts and observations");
@@ -341,7 +354,7 @@ int main(int argc, char** argv) {
     vector<size_t> obs_id;
     vector<string> config_files;
     int station_start, station_count;
-    bool overwrite, profile, save_tests, convert_wind;
+    bool overwrite, profile, save_tests, unwrap_obs, convert_wind;
     string u_name, v_name, spd_name, dir_name;
     long int ai_flt_radius;
 
@@ -393,6 +406,7 @@ int main(int argc, char** argv) {
             ("save-sims-time-index", bool_switch(&(config.save_sims_time_index))->default_value(config.save_sims_time_index), "[Optional] Save time indices of similarity.")
             ("save-sims-station-index", bool_switch(&(config.save_sims_station_index))->default_value(config.save_sims_station_index), "[Optional] Save station indices of similarity.")
             ("save-tests", bool_switch(&save_tests)->default_value(false), "[Optional] Save test forecasts and observations if available")
+            ("unwrap-test-obs", bool_switch(&unwrap_obs)->default_value(false), "[Optional] When saving test observations (--save-tests), unwrap observation time series to align it with forecasts")
             ("quick-sort", bool_switch(&(config.quick_sort))->default_value(config.quick_sort), "[Optional] Use nth_element sort. Change this in *.cfg ")
             ("convert-wind", bool_switch(&(convert_wind))->default_value(false), "[Optional] Use this option if your forecasts have only wind U and V components and you need to convert them to wind speed and direction. Please also specify --name-u --name-v --name-spd --name-dir. Wind speed and direction values will be calculated internally and replacing U and V components respectively.")
             ("name-u", value<string>(&u_name)->default_value("U"), "[Optional] Parameter name for U component of wind")
@@ -512,7 +526,7 @@ int main(int argc, char** argv) {
 
     runAnEnNcdf(forecast_file, observation_file, station_start, station_count,
             obs_id, test_start, test_end, search_start, search_end, fileout, 
-            algorithm, config, overwrite, profile, save_tests, convert_wind,
+            algorithm, config, overwrite, profile, save_tests, unwrap_obs, convert_wind,
             u_name, v_name, spd_name, dir_name, embedding_model, similarity_model, ai_flt_radius);
 
 #if defined(_USE_MPI_EXTENSION)